Metodo Find del controller FindAircraftController funzionante
Metodo Find funzionante, aggiornato il metodo index per consentire la
visualizzazione corretta.

Modificato l'HTML di prova nella index di FindAircraft per la
visualizzazione della descrizione

+++ Codice/FlightInfo/AirplanesAPI/Models/Ac.cs
@@ -25,7 +25,7 @@ public class Ac
     public int Baro_rate { get; set; }
     public string Squawk { get; set; }
     public string Category { get; set; }
-    public int Nav_qnh { get; set; }
+    public double Nav_qnh { get; set; }
     public int Nav_altitude_mcp { get; set; }
     public int Nav_altitude_fms { get; set; }
     public double Rr_lat { get; set; }
